Rename canva.js to canvas.js
