+++ packages/e2e-test-utils-playwright/src/page/canvas.js
